+++ app/assets/javascripts/kuhsaft/cms/customizations.js
@@ -1,1 +1,0 @@
-// Override this empty file to add custom JS behaviour!